+++ src/main/java/com/uacm/mapeo/presupuestos/model/entity/Cliente.java
@@ -13,7 +13,7 @@ public class Cliente {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_cliente")
-    private Long idCliente;
+    private Integer idCliente;
 
     @Column(name = "nombre", length = 150, nullable = false)
     private String nombre;
