+++ src/main/java/com/uacm/mapeo/presupuestos/model/entity/Contrato.java
@@ -14,7 +14,7 @@ public class Contrato {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_contrato")
-    private Long idContrato;
+    private  Integer idContrato;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "id_cliente", nullable = false)
