+++ src/main/java/com/uacm/mapeo/presupuestos/model/entity/Proyecto.java
@@ -16,7 +16,7 @@ public class Proyecto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_proyecto")
-    private Long idProyecto;
+    private Integer idProyecto;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "id_cliente", nullable = false)
